pyproject.toml
[project]
name = "handbook-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf2>=2.8.7",
    "gradio>=6.12.0",
    "huggingface-hub>=0.35.3",
    "lightrag-hku>=1.4.14",
    "openai>=2.6.0",
    "pdfplumber>=0.11.7",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "supabase>=2.28.3",
]
